Refactor move updateBlacklist to NodeDecorator #836
