+++ visualization/app/codeCharta/util/nodeDecorator.spec.ts
@@ -16,7 +16,7 @@ describe("nodeDecorator", () => {
 
 	beforeEach(() => {
 		file = _.cloneDeep(TEST_DELTA_MAP_A)
-		map = _.cloneDeep(TEST_DELTA_MAP_A.map)
+		map = file.map
 		deltaMap = _.cloneDeep(VALID_NODE_WITH_PATH_AND_DELTAS)
 		metricData = [
 			{ name: "rloc", maxValue: 999999 },
@@ -32,11 +32,12 @@ describe("nodeDecorator", () => {
 			edges: { pairingRate: AttributeTypeValue.relative }
 		}
 		blacklist = _.cloneDeep(STATE.fileSettings.blacklist)
+		NodeDecorator.preDecorateFile(file)
 	})
 
 	describe("decorateMap", () => {
 		it("should aggregate given absolute metrics correctly", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 
 			expect(map.attributes["rloc"]).toBe(200)
@@ -44,7 +45,7 @@ describe("nodeDecorator", () => {
 		})
 
 		it("should aggregate given relative metrics correctly", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 
 			expect(map.attributes["functions"]).toBe(100)
@@ -53,7 +54,7 @@ describe("nodeDecorator", () => {
 		it("should aggregate absolute edge metrics correctly", () => {
 			map.children[0].edgeAttributes = { avgCommits: { incoming: 12, outgoing: 13 } }
 			map.children[1].children[0].edgeAttributes = { avgCommits: { incoming: 10, outgoing: 10 } }
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, edgeMetricData, false, attributeTypes)
 
@@ -64,7 +65,7 @@ describe("nodeDecorator", () => {
 		it("should aggregate given relative edge metrics correctly", () => {
 			map.children[0].edgeAttributes = { pairingRate: { incoming: 12, outgoing: 13 } }
 			map.children[1].children[0].edgeAttributes = { pairingRate: { incoming: 10, outgoing: 10 } }
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, edgeMetricData, false, attributeTypes)
 
@@ -74,7 +75,7 @@ describe("nodeDecorator", () => {
 
 		it("should aggregate missing metrics correctly", () => {
 			metricData.push({ name: "some", maxValue: 999999 })
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 
 			expect(map.attributes["rloc"]).toBe(200)
@@ -84,7 +85,7 @@ describe("nodeDecorator", () => {
 
 		it("leaves should have all metrics", () => {
 			metricData.push({ name: "some", maxValue: 999999 })
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 
 			const h = d3.hierarchy(map)
 			h.leaves().forEach(node => {
@@ -100,7 +101,7 @@ describe("nodeDecorator", () => {
 			map.children[0].attributes = undefined
 			metricData.push({ name: "some", maxValue: 999999 })
 
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			const h = d3.hierarchy(map)
 			h.leaves().forEach(node => {
 				expect(node.data.attributes).toBeDefined()
@@ -137,7 +138,8 @@ describe("nodeDecorator", () => {
 					]
 				}
 			]
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.preDecorateFile(file)
+			NodeDecorator.decorateMap(map, metricData, [])
 			expect(map.name).toBe("root/middle")
 			expect(map.children.length).toBe(2)
 			expect(map.children[0].name).toBe("a")
@@ -172,7 +174,8 @@ describe("nodeDecorator", () => {
 					]
 				}
 			]
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.preDecorateFile(file)
+			NodeDecorator.decorateMap(map, metricData, [])
 			expect(map.link).toBe("link1")
 		})
 
@@ -206,7 +209,7 @@ describe("nodeDecorator", () => {
 					]
 				}
 			]
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			expect(map.path).toBe("/root/middle")
 		})
 
@@ -229,7 +232,8 @@ describe("nodeDecorator", () => {
 					]
 				}
 			]
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.preDecorateFile(file)
+			NodeDecorator.decorateMap(map, metricData, [])
 			expect(map.name).toBe("root/middle")
 			expect(map.children.length).toBe(1)
 			expect(map.children[0].name).toBe("singleLeaf")
@@ -286,7 +290,8 @@ describe("nodeDecorator", () => {
 					]
 				}
 			]
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.preDecorateFile(file)
+			NodeDecorator.decorateMap(map, metricData, [])
 			expect(map.name).toBe("root/start")
 			expect(map.children.length).toBe(2)
 			expect(map.children[0].name).toBe("middle/middle2")
@@ -317,7 +322,7 @@ describe("nodeDecorator", () => {
 			map.children[1].attributes = { some: 1 }
 			metricData.push({ name: "some", maxValue: 999999 })
 
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 			const h = d3.hierarchy(map)
 			h.each(node => {
@@ -327,7 +332,7 @@ describe("nodeDecorator", () => {
 		})
 
 		it("all nodes should have an attribute list with listed and available metrics", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 			const h = d3.hierarchy(map)
 			h.each(node => {
@@ -338,7 +343,7 @@ describe("nodeDecorator", () => {
 		})
 
 		it("folders should have sum attributes of children for absolute metrics", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 			const h = d3.hierarchy(map)
 			expect(h.data.attributes["rloc"]).toBe(200)
@@ -346,14 +351,14 @@ describe("nodeDecorator", () => {
 		})
 
 		it("folders should have median attributes of children for relative metrics", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(map, blacklist, metricData, [], false, attributeTypes)
 			const h = d3.hierarchy(map)
 			expect(h.data.attributes["functions"]).toBe(100)
 		})
 
 		it("folders should have sum delta values of children for absolute metrics", () => {
-			NodeDecorator.decorateMap(deltaMap, metricData)
+			NodeDecorator.decorateMap(deltaMap, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(deltaMap, blacklist, metricData, [], true, attributeTypes)
 			const h = d3.hierarchy(deltaMap)
 			expect(h.data.deltas["rloc"]).toBe(295)
@@ -362,14 +367,14 @@ describe("nodeDecorator", () => {
 		})
 
 		it("folders should have median delta values of children for relative metrics", () => {
-			NodeDecorator.decorateMap(deltaMap, metricData)
+			NodeDecorator.decorateMap(deltaMap, metricData, [])
 			NodeDecorator.decorateParentNodesWithAggregatedAttributes(deltaMap, blacklist, metricData, [], true, attributeTypes)
 			const h = d3.hierarchy(deltaMap)
 			expect(h.data.deltas["functions"]).toBe(-3)
 		})
 
 		it("maps with no attribute nodes should be accepted and an attributes member added", () => {
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 
 			const h = d3.hierarchy(map)
 
@@ -380,7 +385,7 @@ describe("nodeDecorator", () => {
 
 		it("all nodes should have a unary attribute", () => {
 			map.children[0].attributes = {}
-			NodeDecorator.decorateMap(map, metricData)
+			NodeDecorator.decorateMap(map, metricData, [])
 			const h = d3.hierarchy(map)
 			h.each(node => {
 				expect(node.data.attributes[MetricService.UNARY_METRIC]).toBeDefined()
